third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/powerfulFeatureRestrictions/geolocation-on-insecure-origin-expected.txt

CONSOLE WARNING: line 14: getCurrentPosition() and watchPosition() no longer work on insecure origins. To use this feature, you should consider switching your application to a secure origin, such as HTTPS. See https://goo.gl/rStTGz for more details.
This is a testharness.js-based test.
PASS getCurrentPosition 
PASS watchPosition 
Harness: the test ran to completion.
